BUILD: listener: fix build warning on global_listener_rwlock without threads
The global_listener_rwlock was introduced by recent commit 13e86d947
("BUG/MEDIUM: listener: Fix race condition when updating the global mngmt
task"), but it's declared static and is not used when threads are disabled,
thus causing a warning to be emitted in this case. Let's just condition it
to thread usage to shut the warning.

This will need to be backported where the patch above is backported.

(cherry picked from commit 469fa479501f4807d9983ca46618aba3c4ec8cb7)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>