MINOR: h1: Consider empty port as invalid in authority for CONNECT
For now, this change is useless because http_get_host_port() returns
IST_NULL when the port is empty. But this will change. For other methods,
empty ports are valid. But not for CONNECT method. To still return a
400-Bad-Request if a CONNECT is performed with an empty port, istlen() is
used to test the port, instead of isttest().

(cherry picked from commit 75348c2e8b7779d1b00d849ecb5b8324033210f7)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>