DOC: config: mention that a single monitor-uri rule is supported
It was reported in issue #1059 that when multiple monitor-uri rules are
specified, only the last one is used. While this was done on purpose
since a single URI is used, it was not clearly mentioned in the doc,
possibly leading to confusion or wasted time trying to establish a
working setup. Let's clarify this point.

(cherry picked from commit 7fe0c62516ac0cb54290f8b6febe351038792e98)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>