SCRIPTS: announce-release: add a link to the data plane API
Since Marko announced at HAProxyConf 2022 that the data plane API is
mostly complete and will now follow the same release cycle as haproxy
starting with 2.7, it's probably the right moment to encourage users
to start trying it so that we can hope to migrate all the painful
discovery stuff there in a not too distant future.

Let's just point to the latest release for now. We'll see in the future
if we need to adapt the link depending on the branch.

(cherry picked from commit e3a02d5e08e4ee4f70bb4553694bc32adb7931fc)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>