MINOR: quic: test IP_PKTINFO support for quic-conn owned socket
Extend the startup platform detection support test for quic-conn owned
socket. It is required to be able to retrieve destination address on a
recvfrom() system call so check if IP_PKTINFO or IP_RECVDSTADDR flags
are supported.

This change is part of quic-conn owned socket implementation.
It may be backported to 2.7 after a period of observation.

(cherry picked from commit 8d46acdfcb49b4132f4a2ab7313315ca54e30ba0)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>