Revert "BUG/MINOR: http-htx: Don't consider an URI as normalized after a set-uri action"
This reverts commit 7533b98b8a4023f815eb7374fa3118bed67a3fc5.

This fix is reverted for now because it may introduce issues with some
config. So we want to announce the possible breakage in the 2.6.7 and
include the fix into the 2.6.8. This way, users will have some time to
modify their configuration.

(cherry picked from commit f7acbd84078ddce63f1f484531cce1f2ac8718da)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>