BUG/MEDIUM: freq-ctr: Don't compute overshoot value for empty counters
The function computing the excess of events of a frequency counter over the
current period for a target frequency must handle empty counters (no event
and no start date). In this case, no excess must be reported.

Because of this bug, long pauses may be experienced on the bandwith
limitation filter.

This patch must be backported to 2.7.

(cherry picked from commit 427293420fe5358539a395a2a843c8707582871d)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>